refactor: use event-delegation for banana click
