+++ public/src.5881f0df.css
@@ -1,6 +1,159 @@
+:root {
+  --animation-play-speed: 8000ms;
+  --animation-play-state: paused;
+  --cubic-bezier-standard: cubic-bezier(0.5, 0, 0.1, 1);
+  --white-50: rgba(255, 255, 255, 0.5);
+}
+
 * {
   box-sizing: border-box;
   margin: 0;
   padding: 0;
   font-family: sans-serif;
+  color: white;
+}
+
+main {
+  overflow: hidden;
+}
+
+button {
+  appearance: none;
+  background: none;
+  border: none;
+}
+
+button > * {
+  pointer-events: none;
+}
+
+.btn {
+  width: 100px;
+  border: 2px solid white;
+  color: white;
+  height: 40px;
+  margin: 0 1rem;
+  font-size: 1rem;
+}
+
+.header {
+  display: flex;
+  background-color: palevioletred;
+  width: 100%;
+  justify-content: center;
+  align-items: center;
+  position: relative;
+  height: 5vh;
+  z-index: 3;
+}
+
+.grid {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  background: paleturquoise;
+  height: 95vh;
+  overflow: hidden;
+}
+
+.column {
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  top: -100px;
+  align-items: center;
+}
+
+@keyframes slideDown {
+  0% {
+    transform: translateY(0);
+  }
+
+  100% {
+    transform: translateY(100vh);
+  }
+}
+
+@keyframes swing {
+  0% {
+    transform: rotate(5deg);
+  }
+  100% {
+    transform: rotate(-5deg);
+  }
+}
+
+@keyframes exit {
+  0% {
+    transform: scale(1);
+  }
+
+  4%,
+  10% {
+    transform: scale(0.5);
+  }
+
+  25%,
+  40% {
+    transform: scale(1.25);
+  }
+
+  50%,
+  53% {
+    transform: rotate(50deg);
+  }
+
+  75% {
+    opacity: 1;
+  }
+
+  100% {
+    transform: rotate(-150deg) scale(0.1);
+    opacity: 0;
+  }
+}
+
+.slideDown-animation {
+  animation: slideDown var(--animation-play-speed) linear;
+  animation-play-state: var(--animation-play-state);
+}
+
+.swing-animation {
+  animation: swing ease-in-out 1s infinite alternate;
+  transform-origin: center -20px;
+}
+
+.exit-animation {
+  animation: exit 750ms var(--cubic-bezier-standard);
+}
+
+.banana {
+  position: absolute;
+  top: 0;
+  display: inline-flex;
+  justify-content: center;
+  align-items: center;
+  width: 2rem;
+  height: 2rem;
+  font-size: 2rem;
+}
+
+.banana > span:after {
+  content: attr(data-points) "pts";
+  font-size: 0.875rem;
+  position: absolute;
+  top: 50%;
+  left: -40%;
+  background-color: var(--white-50);
+  padding: 2px 10px;
+  color: rgba(1, 1, 1, 1);
+  border-radius: 4px;
+}
+
+.banana:focus {
+  outline: none;
+}
+
+.banana:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
 }
